big ol changes
added a player class which has functions and data associated with the player

split game.py into many other files
