+++ __init__.py
@@ -1,0 +1,1 @@
+# initialize the game package
