+++ places/__init__.py
@@ -1,0 +1,1 @@
+# initialize places package
